requirements.txt
flask_migrate
flask_script
flask_sqlalchemy

mysql-python